+++ Tests/BasicCacheTests.swift
@@ -46,27 +46,6 @@ class BasicCacheTests: QuickSpec {
         )
       }
       
-      context("when calling perform") {
-        let key = "key to test"
-        var request: Future<Int>!
-        
-        beforeEach {
-          request = cache.perform(key)
-        }
-        
-        it("should call the closure") {
-          expect(numberOfTimesCalledGet).to(equal(1))
-        }
-        
-        it("should pass the right key") {
-          expect(didGetKey).to(equal(key))
-        }
-        
-        it("should not modify the request") {
-          expect(request).to(beIdenticalTo(fakeRequest.future))
-        }
-      }
-      
       context("when calling get") {
         let key = "key to test"
         var request: Future<Int>!